14/8

+++ api/src/main/java/com/thitracnghiem/api/entities/test/entities/Test.java
@@ -32,5 +32,5 @@ public class Test {
     private Date ngayThi;
 
 
-    private float diem;
+    private Double diem;
 }
